requirements.txt
requests==2.31.0
pandas==2.2.2
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
apache-airflow==2.9.1
dbt-postgres==1.8.0
pytest==8.2.0
